HZN-1564: Modify EventSinkIT to send events from Minion
