--- smoke-test/src/test/java/org/opennms/smoketest/EventSinkIT.java
@@ -1,143 +1,0 @@
-/*******************************************************************************
- * This file is part of OpenNMS(R).
- *
- * Copyright (C) 2018 The OpenNMS Group, Inc.
- * OpenNMS(R) is Copyright (C) 1999-2018 The OpenNMS Group, Inc.
- *
- * OpenNMS(R) is a registered trademark of The OpenNMS Group, Inc.
- *
- * OpenNMS(R) is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published
- * by the Free Software Foundation, either version 3 of the License,
- * or (at your option) any later version.
- *
- * OpenNMS(R) is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with OpenNMS(R).  If not, see:
- *      http://www.gnu.org/licenses/
- *
- * For more information contact:
- *     OpenNMS(R) Licensing <license@opennms.org>
- *     http://www.opennms.org/
- *     http://www.opennms.com/
- *******************************************************************************/
-
-package org.opennms.smoketest;
-
-import static com.jayway.awaitility.Awaitility.await;
-import static java.util.concurrent.TimeUnit.MINUTES;
-import static java.util.concurrent.TimeUnit.SECONDS;
-import static org.hamcrest.Matchers.notNullValue;
-import static org.junit.Assert.assertNotNull;
-
-import java.net.InetAddress;
-import java.net.InetSocketAddress;
-import java.net.UnknownHostException;
-import java.nio.charset.StandardCharsets;
-import java.util.Date;
-import java.util.Properties;
-
-import org.apache.kafka.clients.producer.KafkaProducer;
-import org.apache.kafka.clients.producer.ProducerConfig;
-import org.apache.kafka.clients.producer.ProducerRecord;
-import org.apache.kafka.common.serialization.ByteArraySerializer;
-import org.apache.kafka.common.serialization.StringSerializer;
-import org.junit.Assume;
-import org.junit.Before;
-import org.junit.ClassRule;
-import org.junit.Test;
-import org.opennms.core.camel.JmsQueueNameFactory;
-import org.opennms.core.criteria.CriteriaBuilder;
-import org.opennms.core.ipc.sink.model.SinkMessageProtos;
-import org.opennms.core.xml.JaxbUtils;
-import org.opennms.netmgt.dao.api.NodeDao;
-import org.opennms.netmgt.dao.hibernate.NodeDaoHibernate;
-import org.opennms.netmgt.events.api.EventConstants;
-import org.opennms.netmgt.model.OnmsNode;
-import org.opennms.netmgt.model.events.EventBuilder;
-import org.opennms.netmgt.xml.event.Log;
-import org.opennms.smoketest.utils.DaoUtils;
-import org.opennms.smoketest.utils.HibernateDaoFactory;
-import org.opennms.test.system.api.NewTestEnvironment.ContainerAlias;
-import org.opennms.test.system.api.TestEnvironment;
-import org.opennms.test.system.api.TestEnvironmentBuilder;
-
-import com.google.protobuf.ByteString;
-
-public class EventSinkIT {
-
-    private static final String SENDER_IP = "192.168.1.1";
-    protected static TestEnvironment m_testEnvironment;
-
-    @ClassRule
-    public static final TestEnvironment getTestEnvironment() {
-        if (!OpenNMSSeleniumTestCase.isDockerEnabled()) {
-            return new NullTestEnvironment();
-        }
-        try {
-            final TestEnvironmentBuilder builder = TestEnvironment.builder().opennms().kafka();
-            builder.withOpenNMSEnvironment().addFile(
-                    EventSinkIT.class.getResource("/opennms.properties.d/event-sink-kafka.properties"),
-                    "etc/opennms.properties.d/event-sink-kafka.properties");
-            OpenNMSSeleniumTestCase.configureTestEnvironment(builder);
-            m_testEnvironment = builder.build();
-            return m_testEnvironment;
-        } catch (final Throwable t) {
-            throw new RuntimeException(t);
-        }
-    }
-
-    @Before
-    public void checkForDockerAndLoadExecutor() {
-        Assume.assumeTrue(OpenNMSSeleniumTestCase.isDockerEnabled());
-    }
-
-    @Test
-    public void testEventSinkWithKafka() throws UnknownHostException {
-        Date startOfTest = new Date();
-        String kafkaHost = m_testEnvironment.getContainerInfo(ContainerAlias.KAFKA).networkSettings().ipAddress();
-        kafkaHost = kafkaHost + ":9092";
-        EventBuilder eventBuilder = new EventBuilder(EventConstants.NEW_SUSPECT_INTERFACE_EVENT_UEI, "system-test");
-        eventBuilder.setHost(SENDER_IP);
-        eventBuilder.setInterface(InetAddress.getByName(SENDER_IP));
-        Log eventLog = new Log();
-        eventLog.addEvent(eventBuilder.getEvent());
-        String xmlString = JaxbUtils.marshal(eventLog);
-
-        final Properties producerConfig = new Properties();
-        producerConfig.put(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, kafkaHost);
-        producerConfig.put(ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG, StringSerializer.class.getCanonicalName());
-        producerConfig.put(ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG, ByteArraySerializer.class.getCanonicalName());
-        KafkaProducer<String, byte[]> producer = new KafkaProducer<>(producerConfig);
-        final JmsQueueNameFactory topicNameFactory = new JmsQueueNameFactory("Sink", "Events");
-        byte[] sinkMessageInBytes = wrapMessageToProto("1", xmlString.getBytes(StandardCharsets.UTF_8));
-        ProducerRecord<String, byte[]> producerRecord = new ProducerRecord<String, byte[]>(topicNameFactory.getName(),
-                sinkMessageInBytes);
-        producer.send(producerRecord);
-
-        InetSocketAddress pgsql = m_testEnvironment.getServiceAddress(ContainerAlias.POSTGRES, 5432);
-        HibernateDaoFactory daoFactory = new HibernateDaoFactory(pgsql);
-        NodeDao nodeDao = daoFactory.getDao(NodeDaoHibernate.class);
-
-        final OnmsNode onmsNode = await().atMost(1, MINUTES).pollInterval(5, SECONDS)
-                .until(DaoUtils.findMatchingCallable(nodeDao, new CriteriaBuilder(OnmsNode.class).eq("label", SENDER_IP)
-                        .ge("createTime", startOfTest).toCriteria()), notNullValue());
-
-        assertNotNull(onmsNode);
-
-        producer.close();
-    }
-
-    private byte[] wrapMessageToProto(String messageId, byte[] messageInBytes) {
-        SinkMessageProtos.SinkMessage sinkMessage = SinkMessageProtos.SinkMessage.newBuilder()
-                .setMessageId(messageId)
-                .setContent(ByteString.copyFrom(messageInBytes))
-                .build();
-        return sinkMessage.toByteArray();
-    }
-
-}
