+++ smoke-test/src/test/resources/featuresBoot.d/kafka-rpc.boot
@@ -1,3 +1,5 @@
 #Disable JMS and enable Kafka
 !opennms-core-ipc-rpc-jms
+!minion-jms
 opennms-core-ipc-rpc-kafka
+
