+++ smoke-test/src/test/resources/featuresBoot.d/kafka.boot
@@ -1,4 +1,5 @@
 # Disable installation of the JMS sink
 !opennms-core-ipc-sink-camel
+!minion-jms
 # Install the Apache Kafka sink 
 opennms-core-ipc-sink-kafka
