--- smoke-test/src/test/resources/opennms.properties.d/event-sink-kafka.properties
@@ -1,2 +1,0 @@
-# Switch to the Apache Kafka sink strategy
-org.opennms.core.ipc.sink.strategy=kafka
